backend/requirements/common.txt
celery==4.3.0
redis==3.2.1
Flask==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.3.1
gunicorn==19.9.0
pika==0.12.0
psycopg2==2.7.6
requests==2.20.1
retry==0.9.2
sentry-sdk==0.7.2
boto3==1.9.86
google-cloud-storage==1.14.0
pyjwt==1.7.1
Authlib==0.10
